[ENH] : suppress useless SQL queries
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@14176 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/tikilib.php
@@ -3153,9 +3153,11 @@ function add_pageview() {
 	$this->query($query, $bindvars, -1, -1, false);
 	$query = "insert into `tiki_sessions`(`sessionId`,`timestamp`,`user`,`tikihost`) values(?,?,?,?)";
 	$result = $this->query($query, array($sessionId, (int)$this->now, $user,$_SERVER['HTTP_HOST']), -1, -1, false );
-	// clean up adodb sessions as well in case adodb session garbage collection not working
-	$query = "delete from `sessions` where `expiry`<?";
-	$this->query($query, array($oldy));
+	if ($prefs['session_db'] == 'y') {
+		// clean up adodb sessions as well in case adodb session garbage collection not working
+		$query = "delete from `sessions` where `expiry`<?";
+		$this->query($query, array($oldy));
+	}
 	return true;
     }
